refactor(lockfile): extract inline test modules into dedicated files
Move five inline `#[cfg(test)] mod tests { ... }` blocks into external
`tests.rs` files following CODE_STYLE_GUIDE.md: `resolution` (162
lines), `pkg_ver_peer` (129), `snapshot_dep_ref` (99),
`save_lockfile` (99), `project_snapshot` (84). No behavioral change.

Part of #358.
