+++ crates/lockfile/src/resolution.rs
@@ -97,164 +97,4 @@ impl From<LockfileResolution> for ResolutionSerde {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::serialize_yaml;
-    use pretty_assertions::assert_eq;
-    use text_block_macros::text_block;
-
-    fn integrity(integrity_str: &str) -> Integrity {
-        integrity_str.parse().expect("parse integrity string")
-    }
-
-    #[test]
-    fn deserialize_tarball_resolution() {
-        eprintln!("CASE: without integrity");
-        let yaml = text_block! {
-            "tarball: file:ts-pipe-compose-0.2.1.tgz"
-        };
-        let received: LockfileResolution = serde_saphyr::from_str(yaml).unwrap();
-        dbg!(&received);
-        let expected = LockfileResolution::Tarball(TarballResolution {
-            tarball: "file:ts-pipe-compose-0.2.1.tgz".to_string(),
-            integrity: None,
-        });
-        assert_eq!(received, expected);
-
-        eprintln!("CASE: with integrity");
-        let yaml = text_block! {
-            "tarball: file:ts-pipe-compose-0.2.1.tgz"
-            "integrity: sha512-gf6ZldcfCDyNXPRiW3lQjEP1Z9rrUM/4Cn7BZbv3SdTA82zxWRP8OmLwvGR974uuENhGCFgFdN11z3n1Ofpprg=="
-        };
-        let received: LockfileResolution = serde_saphyr::from_str(yaml).unwrap();
-        dbg!(&received);
-        let expected = LockfileResolution::Tarball(TarballResolution {
-            tarball: "file:ts-pipe-compose-0.2.1.tgz".to_string(),
-            integrity: integrity("sha512-gf6ZldcfCDyNXPRiW3lQjEP1Z9rrUM/4Cn7BZbv3SdTA82zxWRP8OmLwvGR974uuENhGCFgFdN11z3n1Ofpprg==").into()
-        });
-        assert_eq!(received, expected);
-    }
-
-    #[test]
-    fn serialize_tarball_resolution() {
-        eprintln!("CASE: without integrity");
-        let resolution = LockfileResolution::Tarball(TarballResolution {
-            tarball: "file:ts-pipe-compose-0.2.1.tgz".to_string(),
-            integrity: None,
-        });
-        let received = serialize_yaml::to_string(&resolution).unwrap();
-        let received = received.trim();
-        eprintln!("RECEIVED:\n{received}");
-        let expected = text_block! {
-            "tarball: file:ts-pipe-compose-0.2.1.tgz"
-        };
-        assert_eq!(received, expected);
-
-        eprintln!("CASE: with integrity");
-        let resolution = LockfileResolution::Tarball(TarballResolution {
-            tarball: "file:ts-pipe-compose-0.2.1.tgz".to_string(),
-            integrity: integrity("sha512-gf6ZldcfCDyNXPRiW3lQjEP1Z9rrUM/4Cn7BZbv3SdTA82zxWRP8OmLwvGR974uuENhGCFgFdN11z3n1Ofpprg==").into()
-        });
-        let received = serialize_yaml::to_string(&resolution).unwrap();
-        let received = received.trim();
-        eprintln!("RECEIVED:\n{received}");
-        let expected = text_block! {
-            "tarball: file:ts-pipe-compose-0.2.1.tgz"
-            "integrity: sha512-gf6ZldcfCDyNXPRiW3lQjEP1Z9rrUM/4Cn7BZbv3SdTA82zxWRP8OmLwvGR974uuENhGCFgFdN11z3n1Ofpprg=="
-        };
-        assert_eq!(received, expected);
-    }
-
-    #[test]
-    fn deserialize_registry_resolution() {
-        let yaml = text_block! {
-            "integrity: sha512-gf6ZldcfCDyNXPRiW3lQjEP1Z9rrUM/4Cn7BZbv3SdTA82zxWRP8OmLwvGR974uuENhGCFgFdN11z3n1Ofpprg=="
-        };
-        let received: LockfileResolution = serde_saphyr::from_str(yaml).unwrap();
-        dbg!(&received);
-        let expected = LockfileResolution::Registry(RegistryResolution {
-            integrity: integrity(
-                "sha512-gf6ZldcfCDyNXPRiW3lQjEP1Z9rrUM/4Cn7BZbv3SdTA82zxWRP8OmLwvGR974uuENhGCFgFdN11z3n1Ofpprg==",
-            ),
-        });
-        assert_eq!(received, expected);
-    }
-
-    #[test]
-    fn serialize_registry_resolution() {
-        let resolution = LockfileResolution::Registry(RegistryResolution {
-            integrity: integrity(
-                "sha512-gf6ZldcfCDyNXPRiW3lQjEP1Z9rrUM/4Cn7BZbv3SdTA82zxWRP8OmLwvGR974uuENhGCFgFdN11z3n1Ofpprg==",
-            ),
-        });
-        let received = serialize_yaml::to_string(&resolution).unwrap();
-        let received = received.trim();
-        eprintln!("RECEIVED:\n{received}");
-        let expected = text_block! {
-            "integrity: sha512-gf6ZldcfCDyNXPRiW3lQjEP1Z9rrUM/4Cn7BZbv3SdTA82zxWRP8OmLwvGR974uuENhGCFgFdN11z3n1Ofpprg=="
-        };
-        assert_eq!(received, expected);
-    }
-
-    #[test]
-    fn deserialize_directory_resolution() {
-        let yaml = text_block! {
-            "type: directory"
-            "directory: ts-pipe-compose-0.2.1/package"
-        };
-        let received: LockfileResolution = serde_saphyr::from_str(yaml).unwrap();
-        dbg!(&received);
-        let expected = LockfileResolution::Directory(DirectoryResolution {
-            directory: "ts-pipe-compose-0.2.1/package".to_string(),
-        });
-        assert_eq!(received, expected);
-    }
-
-    #[test]
-    fn serialize_directory_resolution() {
-        let resolution = LockfileResolution::Directory(DirectoryResolution {
-            directory: "ts-pipe-compose-0.2.1/package".to_string(),
-        });
-        let received = serialize_yaml::to_string(&resolution).unwrap();
-        let received = received.trim();
-        eprintln!("RECEIVED:\n{received}");
-        let expected = text_block! {
-            "type: directory"
-            "directory: ts-pipe-compose-0.2.1/package"
-        };
-        assert_eq!(received, expected);
-    }
-
-    #[test]
-    fn deserialize_git_resolution() {
-        let yaml = text_block! {
-            "type: git"
-            "repo: https://github.com/ksxnodemodules/ts-pipe-compose.git"
-            "commit: e63c09e460269b0c535e4c34debf69bb91d57b22"
-        };
-        let received: LockfileResolution = serde_saphyr::from_str(yaml).unwrap();
-        dbg!(&received);
-        let expected = LockfileResolution::Git(GitResolution {
-            repo: "https://github.com/ksxnodemodules/ts-pipe-compose.git".to_string(),
-            commit: "e63c09e460269b0c535e4c34debf69bb91d57b22".to_string(),
-        });
-        assert_eq!(received, expected);
-    }
-
-    #[test]
-    fn serialize_git_resolution() {
-        let resolution = LockfileResolution::Git(GitResolution {
-            repo: "https://github.com/ksxnodemodules/ts-pipe-compose.git".to_string(),
-            commit: "e63c09e460269b0c535e4c34debf69bb91d57b22".to_string(),
-        });
-        let received = serialize_yaml::to_string(&resolution).unwrap();
-        let received = received.trim();
-        eprintln!("RECEIVED:\n{received}");
-        let expected = text_block! {
-            "type: git"
-            "repo: https://github.com/ksxnodemodules/ts-pipe-compose.git"
-            "commit: e63c09e460269b0c535e4c34debf69bb91d57b22"
-        };
-        assert_eq!(received, expected);
-    }
-}
+mod tests;
