+++ crates/lockfile/src/snapshot_dep_ref.rs
@@ -129,101 +129,4 @@ impl From<PkgVerPeer> for SnapshotDepRef {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use pretty_assertions::assert_eq;
-
-    fn pkg_name(s: &str) -> PkgName {
-        PkgName::parse(s).unwrap()
-    }
-
-    fn ver_peer(s: &str) -> PkgVerPeer {
-        s.parse().unwrap()
-    }
-
-    fn key(s: &str) -> PkgNameVerPeer {
-        s.parse().unwrap()
-    }
-
-    #[test]
-    fn parse_plain_version() {
-        let dep: SnapshotDepRef = "5.1.2".parse().unwrap();
-        assert_eq!(dep, SnapshotDepRef::Plain(ver_peer("5.1.2")));
-    }
-
-    #[test]
-    fn parse_plain_with_peer_suffix() {
-        let dep: SnapshotDepRef = "17.0.2(react@17.0.2)".parse().unwrap();
-        assert_eq!(dep, SnapshotDepRef::Plain(ver_peer("17.0.2(react@17.0.2)")));
-    }
-
-    #[test]
-    fn parse_alias_unscoped_target() {
-        let dep: SnapshotDepRef = "string-width@4.2.3".parse().unwrap();
-        assert_eq!(dep, SnapshotDepRef::Alias(key("string-width@4.2.3")));
-    }
-
-    #[test]
-    fn parse_alias_scoped_target() {
-        let dep: SnapshotDepRef = "@types/react@17.0.49".parse().unwrap();
-        assert_eq!(dep, SnapshotDepRef::Alias(key("@types/react@17.0.49")));
-    }
-
-    #[test]
-    fn parse_alias_with_peer_suffix() {
-        let dep: SnapshotDepRef = "react-dom@17.0.2(react@17.0.2)".parse().unwrap();
-        assert_eq!(dep, SnapshotDepRef::Alias(key("react-dom@17.0.2(react@17.0.2)")));
-    }
-
-    #[test]
-    fn resolve_plain_uses_alias_key_as_target_name() {
-        let dep: SnapshotDepRef = "5.1.2".parse().unwrap();
-        let resolved = dep.resolve(&pkg_name("string-width"));
-        assert_eq!(resolved.to_string(), "string-width@5.1.2");
-    }
-
-    #[test]
-    fn resolve_alias_uses_alias_target_name_not_key() {
-        let dep: SnapshotDepRef = "string-width@4.2.3".parse().unwrap();
-        let resolved = dep.resolve(&pkg_name("string-width-cjs"));
-        assert_eq!(resolved.to_string(), "string-width@4.2.3");
-    }
-
-    #[test]
-    fn display_roundtrip() {
-        for input in [
-            "5.1.2",
-            "17.0.2(react@17.0.2)",
-            "string-width@4.2.3",
-            "@types/react@17.0.49",
-            "react-dom@17.0.2(react@17.0.2)",
-        ] {
-            let dep: SnapshotDepRef = input.parse().unwrap();
-            assert_eq!(dep.to_string(), input);
-        }
-    }
-
-    #[test]
-    fn deserialize_ok() {
-        for (yaml, expected) in [
-            ("5.1.2", "5.1.2"),
-            ("string-width@4.2.3", "string-width@4.2.3"),
-            ("\"17.0.2(react@17.0.2)\"", "17.0.2(react@17.0.2)"),
-        ] {
-            let dep: SnapshotDepRef = serde_saphyr::from_str(yaml).unwrap();
-            assert_eq!(dep.to_string(), expected);
-        }
-    }
-
-    #[test]
-    fn looks_like_alias_rules() {
-        assert!(!looks_like_alias("5.1.2"));
-        assert!(!looks_like_alias("17.0.2(react@17.0.2)"));
-        assert!(looks_like_alias("string-width@4.2.3"));
-        assert!(looks_like_alias("@types/react@17.0.49"));
-        assert!(looks_like_alias("react-dom@17.0.2(react@17.0.2)"));
-        // protocol-like refs are not aliases
-        assert!(!looks_like_alias("link:../foo"));
-        assert!(!looks_like_alias("workspace:*"));
-    }
-}
+mod tests;
